Move migrate blueprint logic into a command
